requirements.txt
evdev
pyautogui
python-imageseach-drov0
termcolor
pyydotool